requirements.txt
colorama==0.4.4
keras==2.7.0
numpy==1.21.4
Pillow==8.4.0
pyfiglet==0.8.post1
requests==2.26.0
selenium==4.0.0
